Move Reporter to common.
