+++ OpenSim/Common/Reporter.cpp
@@ -1,0 +1,71 @@
+/* -------------------------------------------------------------------------- *
+ *                          OpenSim:  Reporter.cpp                            *
+ * -------------------------------------------------------------------------- *
+ * The OpenSim API is a toolkit for musculoskeletal modeling and simulation.  *
+ * See http://opensim.stanford.edu and the NOTICE file for more information.  *
+ * OpenSim is developed at Stanford University and supported by the US        *
+ * National Institutes of Health (U54 GM072970, R24 HD065690) and by DARPA    *
+ * through the Warrior Web program.                                           *
+ *                                                                            *
+ * Copyright (c) 2005-2016 Stanford University and the Authors                *
+ * Author(s): Ajay Seth                                                       *
+ *                                                                            *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may    *
+ * not use this file except in compliance with the License. You may obtain a  *
+ * copy of the License at http://www.apache.org/licenses/LICENSE-2.0.         *
+ *                                                                            *
+ * Unless required by applicable law or agreed to in writing, software        *
+ * distributed under the License is distributed on an "AS IS" BASIS,          *
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.   *
+ * See the License for the specific language governing permissions and        *
+ * limitations under the License.                                             *
+ * -------------------------------------------------------------------------- */
+
+#include "Reporter.h"
+#include <OpenSim/Common/TimeSeriesTable.h>
+
+
+using namespace SimTK;
+
+namespace OpenSim {
+
+//=============================================================================
+// CONSTRUCTOR(S) AND DESTRUCTOR
+//=============================================================================
+//_____________________________________________________________________________
+// Default constructor.
+Reporter::Reporter()
+{
+    setNull();
+    constructProperties();
+}
+
+//=============================================================================
+// CONSTRUCTION METHODS
+//=============================================================================
+
+//_____________________________________________________________________________
+// Set the data members of this Reporter to their null values.
+void Reporter::setNull()
+{
+    setAuthors("Ajay Seth");
+}
+
+//_____________________________________________________________________________
+// Define properties.
+void Reporter::constructProperties()
+{
+    constructProperty_isDisabled(false);
+    constructProperty_report_time_iterval(0.0);
+}
+
+// Create an underlying SimTK::Reporter to represent the OpenSim::Reporter in the 
+// computational system.  Create a SimTK::Reporter::Custom by default.
+void Reporter::extendAddToSystem(SimTK::MultibodySystem& system) const
+{
+    Super::extendAddToSystem(system);
+}
+
+
+
+} // end of namespace OpenSim
